add delete button

+++ src/components/Blog.jsx
@@ -8,22 +8,20 @@ const Blog = ({ blog, blogs, user, setBlogs, setInfoMessage, setErrorMessage }) 
     e.preventDefault();
     console.log("User :" + JSON.stringify(user))
     console.log("Id :" + id)
-    try {
-      await blogService.deleteBlog(id, user)
-      setBlogs(blogs.filter(blog => blog.id!== id))
-      setInfoMessage('Blog "' + blog.title + '" is deleted successfully')
-      setTimeout(() => {
-        setInfoMessage(null)
-      }, 5000)
-    } catch (exception) {
-      setErrorMessage('wrong blog id')
-      setTimeout(() => {
-        setErrorMessage(null)
-      }, 5000)
-    }
-  }
-  const editBlog = async (e, blog) => {
-
+    if (window.confirm("Do you really want to remove blog?")) {
+      try {
+        await blogService.deleteBlog(id, user)
+        setBlogs(blogs.filter(blog => blog.id!== id))
+        setInfoMessage('Blog "' + blog.title + '" is deleted successfully')
+        setTimeout(() => {
+          setInfoMessage(null)
+        }, 5000)
+      } catch (exception) {
+        setErrorMessage('wrong blog id')
+        setTimeout(() => {
+          setErrorMessage(null)
+        }, 5000)
+    }}
   }
   const showAll = (e) => {
     e.preventDefault();
@@ -32,6 +30,7 @@ const Blog = ({ blog, blogs, user, setBlogs, setInfoMessage, setErrorMessage }) 
   const increaseLikes = async (e) => {
     e.preventDefault();
     console.log("Blogbefore" + JSON.stringify(blog))
+    console.log("User " + JSON.stringify(user))
     const likesincreased = {
       title: blog.title,
       author: blog.author,
@@ -40,10 +39,12 @@ const Blog = ({ blog, blogs, user, setBlogs, setInfoMessage, setErrorMessage }) 
       likes: blog.likes + 1,
       id: blog.id,
     }
-    console.log("Blogafter" + JSON.stringify(likesincreased))
+    console.log("BlogUpdated" + JSON.stringify(likesincreased))
     try {
       await blogService.updateBlog(likesincreased, user)
-      setBlogs(blogs.map(blog => blog.id === likesincreased.id? likesincreased : blog))
+      const updatedBlogs = blogs.map(blog => blog.id === likesincreased.id? likesincreased : blog)
+      //sort by most likes
+      setBlogs(updatedBlogs.toSorted((a, b) => b.likes - a.likes))
     } catch (exception) {
       setErrorMessage(exception)
       setTimeout(() => {
@@ -56,22 +57,27 @@ const Blog = ({ blog, blogs, user, setBlogs, setInfoMessage, setErrorMessage }) 
   return (
     <div className='blog'>
       <div style={showLess}>
-        {blog.title}
+        Title: {blog.title}
         <button onClick={(e) => showAll(e)}>view</button>
       </div>
       <div style={showMore}>
         <div>
-          {blog.title}
+          Title: {blog.title}
           <button onClick={(e) => showAll(e)}>hide</button>
         </div>
-        <div>{blog.url}</div>
-        <div>likes: {blog.likes}
-        <button onClick={(e)=> increaseLikes(e)}>like</button>
+        <div>
+          Url: {blog.url}
         </div>
-        <div>{blog.author}</div>
-      </div> 
-        <button onClick={(e) => editBlog(e, blog)}>Edit Blog</button>
-        <button onClick={(e) => deleteBlog(e, blog.id)}>Delete Blog</button>
+        <div>Likes: {blog.likes}
+          <button onClick={(e)=> increaseLikes(e)}>like</button>
+        </div>
+        <div>
+          Author: {blog.author}
+        </div>
+        {(blog.author === user.name)&&
+          <button onClick={(e) => deleteBlog(e, blog.id)}>Delete Blog</button>
+        }
+      </div>   
     </div>  
   )
 }
